Move context_aware_monkeypatch to conda.testing
