+++ conda/testing/__init__.py
@@ -20,18 +20,21 @@ import uuid
 import warnings
 from contextlib import contextmanager, nullcontext
 from dataclasses import dataclass
+from logging import getLogger
 from os.path import dirname, isfile, join, normpath
 from pathlib import Path
 from subprocess import check_output
 from typing import Iterable, overload
 
 import pytest
-from pytest import CaptureFixture, ExceptionInfo
+from pytest import CaptureFixture, ExceptionInfo, MonkeyPatch
 
 from ..base.context import context, reset_context
 from ..cli.main import init_loggers
 from ..common.compat import on_win
 from ..deprecations import deprecated
+
+log = getLogger(__name__)
 
 
 @deprecated("23.9", "24.3")
@@ -309,3 +312,18 @@ def tmp_env(
 ) -> TmpEnvFixture:
     """Fixture returning TmpEnvFixture instance."""
     yield TmpEnvFixture(path_factory, conda_cli)
+
+
+@pytest.fixture(name="monkeypatch")
+def context_aware_monkeypatch(monkeypatch: MonkeyPatch) -> MonkeyPatch:
+    """A monkeypatch fixture that resets context after each test"""
+    yield monkeypatch
+
+    # reset context if any CONDA_ variables were set/unset
+    if any(
+        obj is os.environ and name.startswith("CONDA_")
+        for obj, name, _ in monkeypatch._setitem
+    ):
+        log.debug("monkeypatch cleanup: undo & reset context")
+        monkeypatch.undo()
+        reset_context()
